add conditionDefinition.d.ts
